+++ src/components/radio-group/radio-group.e2e.ts
@@ -1,7 +1,7 @@
 import { newE2EPage } from "@stencil/core/testing";
 
 describe('Radio Group', () => {
-    it('should navigate to the first not disabled option when pressing tab key', async () => {
+    it('should navigate to the first not disabled radio when pressing tab key', async () => {
         // Given
         const page = await newE2EPage();
         await page.setContent(`
@@ -23,7 +23,7 @@ describe('Radio Group', () => {
         expect(activeElId).toEqual('radio-2');
     });
 
-    it('should not make unselected options focusable when an option is clicked', async () => {
+    it('should not make unselected radios focusable when a radio is clicked', async () => {
         // Given
         const page = await newE2EPage();
         await page.setContent(`
@@ -49,7 +49,7 @@ describe('Radio Group', () => {
         expect(activeElId).toEqual('radio-3');
     });
 
-    it('should checked the first option when pressing space key', async () => {
+    it('should checked the first radio when pressing space key', async () => {
         // Given
         const page = await newE2EPage();
         await page.setContent(`
@@ -62,6 +62,7 @@ describe('Radio Group', () => {
             `);
         const firstTabbableItem = await page.find('wcs-input');
         await firstTabbableItem.focus();
+        await page.keyboard.press('Tab');
 
         const wcsRadioGroup = await page.find('wcs-radio-group');
 
@@ -69,19 +70,19 @@ describe('Radio Group', () => {
         await page.keyboard.press('Space');
 
         // Then
-        const firstRadio = await wcsRadioGroup.find('#radio-1');
-        expect(firstRadio.getProperty('checked')).toBeTruthy();
+        const firstRadio = await wcsRadioGroup.find('#radio-1 >>> input');
+        expect(firstRadio.getAttribute('aria-checked')).toEqual('true')
     });
 
-    it('should navigate to the checked option when pressing tab key', async () => {
+    it('should navigate to the checked radio when pressing tab key', async () => {
         // Given
         const page = await newE2EPage();
         await page.setContent(`
                 <wcs-input></wcs-input>
-                <wcs-radio-group>
+                <wcs-radio-group value="3">
                     <wcs-radio disabled id="radio-1" value="1"></wcs-radio>
                     <wcs-radio id="radio-2" value="2"></wcs-radio>
-                    <wcs-radio checked id="radio-3" value="3"></wcs-radio>
+                    <wcs-radio id="radio-3" value="3"></wcs-radio>
                 </wcs-radio-group>
             `);
         const firstTabbableItem = await page.find('wcs-input');
@@ -96,7 +97,7 @@ describe('Radio Group', () => {
     });
 
     describe('When pressing arrow down key', () => {
-        it('should checked the next option not disabled when we are on the first option and next option is here', async () => {
+        it('should checked the next radio not disabled when we are on the first radio and next radio is here', async () => {
             // Given
             const page = await newE2EPage();
             await page.setContent(`
@@ -108,19 +109,19 @@ describe('Radio Group', () => {
                 `);
 
             const wcsRadioGroup = await page.find('wcs-radio-group');
-            const firstOption = await wcsRadioGroup.find('#radio-1');
-            await firstOption.focus();
+            const firstRadio = await wcsRadioGroup.find('#radio-1');
+            await firstRadio.focus();
 
             // When
             await page.keyboard.press('ArrowDown');
             await page.waitForChanges();
 
             // Then
-            const secondRadio = await wcsRadioGroup.find('#radio-3');
-            expect(secondRadio.getProperty('checked')).toBeTruthy();
+            const secondRadio = await wcsRadioGroup.find('#radio-3 >>> input');
+            expect(secondRadio.getAttribute('aria-checked')).toEqual('true');
         });
 
-        it('should checked the first not disabled option when we are on the last option', async () => {
+        it('should checked the first not disabled radio when we are on the last radio', async () => {
             // Given
             const page = await newE2EPage();
             await page.setContent(`
@@ -132,21 +133,110 @@ describe('Radio Group', () => {
                 `);
 
             const wcsRadioGroup = await page.find('wcs-radio-group');
-            const lastOption = await wcsRadioGroup.find('#radio-3');
-            await lastOption.focus();
+            const lastRadio = await wcsRadioGroup.find('#radio-3');
+            await lastRadio.click();
 
             // When
             await page.keyboard.press('ArrowDown');
             await page.waitForChanges();
 
             // Then
-            const firstRadio = await wcsRadioGroup.find('#radio-2');
-            expect(firstRadio.getProperty('checked')).toBeTruthy();
+            const firstRadio = await wcsRadioGroup.find('#radio-2 >>> input');
+            expect(firstRadio.getAttribute('aria-checked')).toEqual('true');
+        });
+    });
+
+    describe('Events handling', () => {
+        it('should fire wcsRadioClick when we click on an unchecked radio', async () => {
+            // Given
+            const page = await newE2EPage();
+            await page.setContent(`
+            <wcs-radio-group>
+                <wcs-radio id="radio-1" value="1"></wcs-radio>
+                <wcs-radio id="radio-2" value="2"></wcs-radio>
+                <wcs-radio id="radio-3" value="3"></wcs-radio>
+            </wcs-radio-group>
+        `);
+
+            const wcsRadioGroup = await page.find('wcs-radio-group');
+            const radio2 = await wcsRadioGroup.find('#radio-2');
+
+            const wcsRadioClick = await page.spyOnEvent('wcsRadioClick');
+
+            // When
+            await radio2.click();
+
+            // Then
+            expect(wcsRadioClick).toHaveReceivedEvent();
+        });
+
+        it('should fire wcsChange when we click or move to an unchecked radio', async () => {
+            // Given
+            const page = await newE2EPage();
+            await page.setContent(`
+            <wcs-radio-group>
+                <wcs-radio id="radio-1" value="1"></wcs-radio>
+                <wcs-radio id="radio-2" value="2"></wcs-radio>
+                <wcs-radio id="radio-3" value="3"></wcs-radio>
+            </wcs-radio-group>
+        `);
+
+            const wcsRadioGroup = await page.find('wcs-radio-group');
+            const radio2 = await wcsRadioGroup.find('#radio-2');
+
+            const wcsChangeEvent = await page.spyOnEvent('wcsChange');
+
+            // When: Clicking on an unchecked radio
+            await radio2.click();
+
+            // Then
+            expect(wcsChangeEvent).toHaveReceivedEvent();
+
+            // When: Moving focus to another unchecked radio with keyboard
+            await radio2.focus();
+            await page.keyboard.press('ArrowDown'); // Move to the next radio
+
+            // Then
+            expect(wcsChangeEvent).toHaveReceivedEventTimes(2);
+        });
+
+        it('should fire wcsBlur and wcsFocus from the radios', async () => {
+            // Given
+            const page = await newE2EPage();
+            await page.setContent(`
+            <wcs-radio-group>
+                <wcs-radio id="radio-1" value="1"></wcs-radio>
+                <wcs-radio id="radio-2" value="2"></wcs-radio>
+                <wcs-radio id="radio-3" value="3"></wcs-radio>
+            </wcs-radio-group>
+        `);
+
+            const radio = await page.find('#radio-1');
+
+            const wcsBlur = await page.spyOnEvent('wcsBlur');
+            const wcsFocus = await page.spyOnEvent('wcsFocus');
+
+            // When: Focusing on the first radio
+            await radio.focus();
+
+            // Then: Expect wcsFocus to be fired
+            expect(wcsFocus).toHaveReceivedEvent();
+
+            // When: Moving to the next radio
+            await page.keyboard.press('ArrowDown');
+            await page.waitForChanges();
+
+            // When: Moving focus to another radio
+            await page.keyboard.press('Tab');
+
+            // Then
+            expect(wcsBlur).toHaveReceivedEvent();
+            expect(wcsFocus).toHaveReceivedEventTimes(2);
         });
     });
 
     describe('When pressing arrow up key', () => {
-        it('should checked the previous option when we are on the last option and previous option is here', async () => {
+        it('should checked the previous radio when we are on the last radio and previous radio is here', async () => {
             // Given
             const page = await newE2EPage();
             await page.setContent(`
@@ -158,19 +248,19 @@ describe('Radio Group', () => {
                 `);
 
             const wcsRadioGroup = await page.find('wcs-radio-group');
-            const lastOption = await wcsRadioGroup.find('#radio-3');
-            await lastOption.focus();
+            const lastRadio = await wcsRadioGroup.find('#radio-3');
+            await lastRadio.focus();
 
             // When
             await page.keyboard.press('ArrowUp');
             await page.waitForChanges();
 
             // Then
-            const secondRadio = await wcsRadioGroup.find('#radio-2');
-            expect(secondRadio.getProperty('checked')).toBeTruthy();
+            const secondRadio = await wcsRadioGroup.find('#radio-2 >>> input');
+            expect(secondRadio.getAttribute('aria-checked')).toEqual('true');
         });
 
-        it('should checked the last option when we are on the first option', async () => {
+        it('should checked the last radio when we are on the first radio', async () => {
             // Given
             const page = await newE2EPage();
             await page.setContent(`
@@ -182,16 +272,16 @@ describe('Radio Group', () => {
                 `);
 
             const wcsRadioGroup = await page.find('wcs-radio-group');
-            const firstOption = await wcsRadioGroup.find('#radio-1');
-            await firstOption.focus();
+            const firstRadio = await wcsRadioGroup.find('#radio-1');
+            await firstRadio.focus();
 
             // When
             await page.keyboard.press('ArrowUp');
             await page.waitForChanges();
 
             // Then
-            const firstRadio = await wcsRadioGroup.find('#radio-3');
-            expect(firstRadio.getProperty('checked')).toBeTruthy();
+            const lastRadio = await wcsRadioGroup.find('#radio-3 >>> input');
+            expect(lastRadio.getAttribute('aria-checked')).toEqual('true');
         });
     });
 });