+++ src/components/radio-group/radio-group.tsx
@@ -12,7 +12,14 @@ import {
 } from '@stencil/core';
 import { RadioGroupChangeEventDetail, RadioGroupMode } from './radio-group-interface';
 import { RadioChosedEvent } from '../radio/radio-interface';
-import { isDownArrowKey, isLeftArrowKey, isRightArrowKey, isUpArrowKey } from "../../utils/helpers";
+import {
+    isDownArrowKey, isEnterKey,
+    isLeftArrowKey,
+    isRightArrowKey,
+    isSpaceKey,
+    isTabKey,
+    isUpArrowKey
+} from "../../utils/helpers";
 import { AriaAttributeName, MutableAriaAttribute } from "../../utils/mutable-aria-attribute";
 
 @Component({
@@ -21,111 +28,124 @@ import { AriaAttributeName, MutableAriaAttribute } from "../../utils/mutable-ari
     shadow: true
 })
 export class RadioGroup implements ComponentInterface, MutableAriaAttribute {
-    @Prop() value: any | any[] | undefined | null;
-    @Prop({reflect: true, mutable: false}) name;
-    @Prop({reflect: true, mutable: false}) mode: RadioGroupMode = 'radio';
     @Element() private el!: HTMLWcsRadioGroupElement;
+    
+    /**
+     * The value of the radio-group. Automatically reflects which radio button is selected.
+     */
+    @Prop({ mutable: true }) value: any | any[] | undefined | null;
+
+    /**
+     * The name of the control to be set on all radio button children
+     */
+    @Prop({ reflect: true }) name;
+
+    /**
+     * The display mode of the control to be set on all radio button children
+     */
+    @Prop({ reflect: true }) mode: RadioGroupMode = 'radio';
 
     /** Emitted when the value has changed. */
     @Event() wcsChange!: EventEmitter<RadioGroupChangeEventDetail>;
 
     @Watch('value')
-    onValueChangeHandler(newValue: any) {
-        this.updateOptionsState(newValue, true);
+    onValueChangeHandler(value: any | undefined) {
+        this.updateRadioTabIndex(value);
+        this.updateAllRadioState();
     }
 
     componentDidLoad() {
-        if (this.value) {
-            this.updateOptionsState(this.value, true);
+       this.onValueChangeHandler(this.value);
+       this.updateAllRadioModeAndName();
+    }
+    
+    onSlotChange() {
+        this.updateRadioTabIndex(this.value);
+        this.updateAllRadioModeAndName();
+    }
+    
+    updateAllRadioState() {
+        this.getSlottedRadios().forEach(r => r.updateState());
+    }
+    
+    @Watch('name')
+    @Watch('mode')
+    updateAllRadioModeAndName() {
+        this.getSlottedRadios().forEach(r => {
+            r.mode = this.mode;
+            r.name = this.name;
+        });
+    }
+    
+    updateRadioTabIndex(value: any | undefined) {
+        const radios = this.getSlottedRadios();
+
+        // Get the first radio that is not disabled and the checked one
+        const first = radios.find((radio) => !radio.disabled);
+        const checked = radios.find((radio) => radio.value === value && !radio.disabled);
+
+        if (!first && !checked) {
+            return;
         }
 
-        // We set the radio group mode on all options
-        this.options.forEach(o => o.mode = this.mode);
+        // If an enabled checked radio exists, set it to be the focusable radio
+        // otherwise we default to focus the first radio
+        const focusable = checked || first;
 
-        this.optionsNotDisabled.forEach((option) => {
-            if (option.checked) {
-                option.tabIndex = 0;
-            } else {
-                option.tabIndex = -1
-            }
-        });
-
-        // If no option is already checked we made the first non-disabled option focusable
-        if (this.optionsNotDisabled.length > 0 && this.optionsNotDisabled.findIndex(o => o.tabIndex === 0) === -1) this.optionsNotDisabled[0].tabIndex = 0;
+        for (const radio of radios) {
+            const tabindex = radio === focusable ? 0 : -1;
+            radio.setTabIndex(tabindex);
+        }
     }
 
-    private get options(): HTMLWcsRadioElement[] {
-        const opts = this.el.querySelectorAll('wcs-radio');
-        const slot = this.el.querySelector('slot');
-        return opts.length !== 0
-            ? opts as unknown as HTMLWcsRadioElement[]
-            : slot !== null
-                ? slot.assignedElements() as HTMLWcsRadioElement[]
-                : [];
+    private getSlottedRadios(): HTMLWcsRadioElement[] {
+        return Array.from(this.el.querySelectorAll('wcs-radio'));
     }
 
-    private get optionsNotDisabled(): HTMLWcsRadioElement[] {
-        return Array.from(this.options).filter(option => !option.disabled);
+    private getSlottedRadiosNotDisabled(): HTMLWcsRadioElement[] {
+        return Array.from(this.getSlottedRadios().filter(radio => !radio.disabled));
     }
 
     @Listen('wcsRadioClick')
-    selectedOptionChanged(event: CustomEvent<RadioChosedEvent>) {
-        this.updateOptionsState(event.detail.value, false);
+    handleRadioClick(event: CustomEvent<RadioChosedEvent>) {
+        this.value = event.detail.value;
         this.wcsChange.emit({
             value: event.detail.value
-        });
-
-        this.optionsNotDisabled.forEach((option) => {
-            if(option === event.detail.source) {
-                option.tabIndex = 0;
-            } else {
-                option.tabIndex = -1;
-            }
         });
     }
 
     @Listen('keydown')
-    handleKeyDown(ev: KeyboardEvent) {
-        if (isDownArrowKey(ev) || isUpArrowKey(ev) || isLeftArrowKey(ev) || isRightArrowKey(ev)) {
+    async handleKeyDown(ev: KeyboardEvent) {
+        const radiosNotDisabled = this.getSlottedRadiosNotDisabled();
+        const previousSelected = radiosNotDisabled.findIndex(r => r === ev.target);
+        let indexToSelect: number;
+
+        if ((isSpaceKey(ev) || isEnterKey(ev))) {
             ev.preventDefault();
-        }
-    }
-
-    @Listen('keyup')
-    async handleKeyUp(ev: KeyboardEvent) {
-        const radiosNotDisabled = this.optionsNotDisabled;
-
-        // Get the index of the currently checked radio
-        const previousSelected = this.optionsNotDisabled.findIndex(o => o.checked) >= 0
-            ? this.optionsNotDisabled.findIndex(o => o.checked)
-            : 0;
-
-        if (isDownArrowKey(ev) || isRightArrowKey(ev)) {
+            indexToSelect = previousSelected;
+        } else if (isDownArrowKey(ev) || isRightArrowKey(ev)) {
+            ev.preventDefault();
             // Check the next wcs-radio from the previous selected
-            const nextOptionIndex = (previousSelected + 1) % radiosNotDisabled.length;  // to return at the beginning on the list when we are on the last index
-            radiosNotDisabled[nextOptionIndex].checked = true;
-            this.wcsChange.emit({
-                value: radiosNotDisabled[nextOptionIndex].value
-            });
+            indexToSelect = (previousSelected + 1) % radiosNotDisabled.length;  // to return at the beginning on the list when we are on the last index
         } else if (isUpArrowKey(ev) || isLeftArrowKey(ev)) {
+            ev.preventDefault();
             // Check the previous wcs-radio from the previous selected
-            const previousOptionIndex = (previousSelected - 1 + radiosNotDisabled.length) % radiosNotDisabled.length; // To return at the end of the list when we are on index=0
-            radiosNotDisabled[previousOptionIndex].checked = true;
-            this.wcsChange.emit({
-                value: radiosNotDisabled[previousOptionIndex].value
-            });
+            indexToSelect = (previousSelected - 1 + radiosNotDisabled.length) % radiosNotDisabled.length; // To return at the end of the list when we are on index=0
+        } else if (isTabKey(ev)) {
+            return;
         }
-    }
-
-    private updateOptionsState(value: string, markAsChecked: boolean) {
-        for (const option of this.options) {
-            if (option.value !== value) {
-                option.removeAttribute('checked');
-            } else {
-                if (markAsChecked) {
-                    option.setAttribute('checked', '');
-                }
+        
+        const radioToSelect = radiosNotDisabled[indexToSelect]
+        if (radioToSelect) {
+            radioToSelect.focus();
+            
+            if (this.value !== radioToSelect.value) {
+                this.wcsChange.emit({
+                    value: radioToSelect.value
+                })
             }
+            
+            this.value = radioToSelect.value;
         }
     }
 
@@ -137,7 +157,7 @@ export class RadioGroup implements ComponentInterface, MutableAriaAttribute {
     render() {
         return (
             <Host role={"radiogroup"}>
-                <slot name="option"/>
+                <slot name="option" onSlotchange={this.onSlotChange.bind(this)}/>
             </Host>
         );
     }