+++ src/components/radio/radio.tsx
@@ -1,18 +1,17 @@
 import {
     Component,
     ComponentInterface,
-    h,
-    Host,
-    Prop,
     Element,
     Event,
     EventEmitter,
-    Listen,
-    Watch
+    h,
+    Host,
+    Method,
+    Prop,
+    State
 } from '@stencil/core';
 import { RadioChosedEvent } from './radio-interface';
 import { RadioGroupMode } from '../radio-group/radio-group-interface';
-import { isEnterKey, isSpaceKey } from '../../utils/helpers';
 
 @Component({
     tag: 'wcs-radio',
@@ -23,51 +22,93 @@ import { isEnterKey, isSpaceKey } from '../../utils/helpers';
 })
 export class Radio implements ComponentInterface {
     private inputId = `wcs-rb-${radioButtonIds++}`;
-    private inputEl: HTMLInputElement;
     @Element() private el!: HTMLWcsRadioElement;
-    @Prop({ reflect: true, mutable: false }) mode: RadioGroupMode = 'radio';
-
-    @Prop({ mutable: true, reflect: true }) value: any | any[] | undefined | null;
-    @Prop({ mutable: true, reflect: true }) label: string;
+    
     /**
      * If `true`, the radio is selected. 
+     * @private
      */
-    @Prop({mutable: true, reflect: true}) checked = false;
+    @State() private checked = false;
+
+    /**
+     * The tabindex of the radio button
+     * @private
+     */
+    @State() private radioTabIndex = -1;
+
+    /**
+     * Sets a unique value for each radio, used to identify which radio button in a group is selected
+     */
+    @Prop({ mutable: true, reflect: true }) value: any | any[] | undefined | null;
+
+    /**
+     * The label text displayed for the user
+     */
+    @Prop({ mutable: true, reflect: true }) label: string;
+    
     /**
      * If `true`, the user cannot interact with the radio.
      */
     @Prop({ mutable: true }) disabled = false;
 
-    @Event({eventName: 'wcsRadioClick' }) wcsRadioClick: EventEmitter<RadioChosedEvent>
+    /**
+     * Emitted when the radio is clicked or Space/Enter is pressed above an unchecked radio
+     */
+    @Event({ eventName: 'wcsRadioClick' }) wcsRadioClick: EventEmitter<RadioChosedEvent>
 
     /**
-     * Giving every radio button the same name is mandatory to group them
+     * Emitted when the radio has focus.
      */
-    @Prop() name: string;
+    @Event() wcsFocus!: EventEmitter<FocusEvent>;
 
-    @Listen('keydown')
-    onKeyDown(_event: KeyboardEvent) {
-        if ((isSpaceKey(_event) || isEnterKey(_event)) && !this.el.checked) {
-            this.checked = true;
-            this.inputEl.click(); // input[radio].checked = true does not trigger any event => input[radio].click() emit a change event
+    /**
+     * Emitted when the radio loses focus.
+     */
+    @Event() wcsBlur!: EventEmitter<FocusEvent>;
+
+    /**
+     * The name of the control, automatically set by the radio group.  
+     * (You shouldn't set this prop by yourself)
+     * @internal
+     */
+    @Prop({ mutable: true }) name: string;
+
+
+    /**
+     * The display mode of the control, automatically set by the radio group.  
+     * (You shouldn't set this prop by yourself)
+     * @internal
+     */
+    @Prop({ reflect: true, mutable: false }) mode: RadioGroupMode = 'radio';
+
+    /**
+     * @internal
+     * Checks of unchecks the radio according to it's radio-group value
+     */
+    @Method()
+    async updateState() {
+        const radioGroup = this.getClosestRadioGroup();
+        if (radioGroup) {
+            this.checked = radioGroup.name === this.name && radioGroup.value === this.value
         }
     }
+
+    /** @internal */
+    @Method()
+    async setTabIndex(value: number) {
+        this.radioTabIndex = value;
+    }
+
+    onFocus(ev: FocusEvent) {
+        this.wcsFocus.emit(ev);
+    }
+
+    onBlur(ev: FocusEvent) {
+        this.wcsBlur.emit(ev);
+    }
     
-    @Watch("checked")
-    checkedChanged(newValue: boolean) {
-        if (newValue) {
-            if (this.el.parentElement.tagName === 'WCS-RADIO-GROUP') {
-                Array.from(this.el.parentElement.querySelectorAll('wcs-radio'))
-                  .filter(radio => radio.name === this.el.name)
-                  .filter(radio => radio !== this.el)
-                  .forEach(radio => {
-                      radio.checked = false;
-                      radio.tabIndex = -1;
-                  });
-                this.el.tabIndex = 0;
-                this.el.focus();
-            }
-        }
+    private getClosestRadioGroup(): HTMLWcsRadioGroupElement | null {
+        return this.el.closest('wcs-radio-group');
     }
 
     componentWillLoad(): Promise<void> | void {
@@ -76,13 +117,17 @@ export class Radio implements ComponentInterface {
             this.value = this.el.innerText || '';
         }
     }
+    
+    private onChange(_: Event) {
+        if (this.disabled) return;
 
-    componentDidLoad() {
-        this.inputEl = this.el.shadowRoot.querySelector('input');
-        this.inputEl.addEventListener('change', _ => {
-            this.emitRadioChangeEvent();
+        // If the radio is unchecked, then the change represents its transition to the check state.
+        // Only emit the change event when going from unchecked to checked, like the native behavior.
+        // https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/change_event
+        if (!this.checked) {
             this.checked = true;
-        })
+            this.emitRadioChangeEvent();
+        }
     }
 
     emitRadioChangeEvent() {
@@ -96,7 +141,9 @@ export class Radio implements ComponentInterface {
     render() {
         
         return (
-            <Host slot="option" {...(this.disabled ? {tabIndex: -1} : {})}>
+            <Host slot="option"
+                  tabIndex={this.disabled ? -1 : this.radioTabIndex}
+                  class={this.checked ? 'checked' : ''}>
                 <input
                     id={this.inputId}
                     type="radio"
@@ -104,6 +151,9 @@ export class Radio implements ComponentInterface {
                     value={this.value}
                     checked={this.checked} // Initial checked state of native input
                     disabled={this.disabled}
+                    onChange={this.onChange.bind(this)}
+                    onFocus={this.onFocus.bind(this)}
+                    onBlur={this.onBlur.bind(this)}
                     aria-disabled={this.disabled ? 'true' : null}
                     aria-checked={`${this.checked}`}
                 />