+++ stories/components/radio-group/radio-group.stories.ts
@@ -1,10 +1,11 @@
 import { Meta } from '@storybook/web-components';
-import { html } from 'lit-html';
+import { html, nothing } from 'lit-html';
 // @ts-ignore
 import { RadioGroupMode } from '../../../src/components/radio-group/radio-group-interface';
 // @ts-ignore
 import { withActions } from '@storybook/addon-actions/decorator';
 import { getComponentArgs } from '../../utils/args-generation';
+import { ifDefined } from "lit-html/directives/if-defined.js";
 
 /**
  * Radio buttons are often used in forms or for other functions that let users select only one option from a list. With 
@@ -18,7 +19,15 @@ import { getComponentArgs } from '../../utils/args-generation';
 const meta: Meta = {
     title: 'Components/Radio-group',
     component: 'wcs-radio-group',
-    argTypes: getComponentArgs('wcs-radio-group'),
+    argTypes: {
+        ...getComponentArgs('wcs-radio-group'),
+        ariaLabel: {
+            description: 'An optional aria-label can be added'
+        },
+        disabled: {
+            description: 'Sets some wcs-radio as disabled'
+        }
+    },
     parameters: {
         actions: {
             handles: [
@@ -36,23 +45,28 @@ export default meta;
 
 interface RadioGroupStoryData {
     disabled: boolean,
-    defaultValue: string,
+    value: string,
     mode: RadioGroupMode,
-    label: string
+    ariaLabel: string,
+    name: string,
 }
 
 export const Default = {
     args: {
         disabled: false,
         mode: 'radio',
-        label: 'Default radio group'
+        ariaLabel: 'Default radio group',
+        name: 'company',
     },
     render: (args: RadioGroupStoryData) => html`
-        <wcs-radio-group value="${args.defaultValue}" aria-label=${args.label} mode=${args.mode}>
-            <wcs-radio label="SNCF" name="company" value="1" ?disabled=${args.disabled}></wcs-radio>
-            <wcs-radio label="SNCF Réseau" name="company" value="2"></wcs-radio>
-            <wcs-radio label="SNCF Voyageurs" name="company" value="3"></wcs-radio>
-            <wcs-radio label="Gares & Connexions" name="company" value="4" ?disabled=${args.disabled}></wcs-radio>
+        <wcs-radio-group name="${ifDefined(args.name)}"
+                         value="${ifDefined(args.value)}"
+                         aria-label=${args.ariaLabel ?? nothing}
+                         mode=${ifDefined(args.mode)}>
+            <wcs-radio label="SNCF" value="1" ?disabled=${args.disabled}></wcs-radio>
+            <wcs-radio label="SNCF Réseau" value="2"></wcs-radio>
+            <wcs-radio label="SNCF Voyageurs" value="3"></wcs-radio>
+            <wcs-radio label="Gares & Connexions" value="4" ?disabled=${args.disabled}></wcs-radio>
         </wcs-radio-group>
     `
 };
@@ -62,7 +76,7 @@ export const Disabled = {
     args: {
         disabled: true,
         mode: 'radio',
-        label: 'Disabled radio group'
+        ariaLabel: 'Disabled radio group',
     },
 };